+++ R/Plot2WayANOVA.R
@@ -42,6 +42,7 @@
 #'                dataframe = NULL,
 #'                confidence=.95,
 #'                plottype = "line",
+#'                errorbar.display = "CI",
 #'                xlab = NULL,
 #'                ylab = NULL,
 #'                title = NULL,
@@ -59,7 +60,11 @@
 #'                overlay.type = NULL,
 #'                posthoc.method = "scheffe",
 #'                show.dots = FALSE,
-#'                PlotSave = FALSE)
+#'                PlotSave = FALSE,
+#'                ggtheme = ggplot2::theme_bw(),
+#'                package = "RColorBrewer",
+#'                palette = "Dark2",
+#'                ggplot.component = NULL)
 #' @param formula a formula with a numeric dependent (outcome) variable,
 #'   and two independent (predictor) variables e.g. \code{mpg ~ am * vs}.
 #'   The independent variables are coerced to factors (with warning) if
@@ -67,6 +72,9 @@
 #' @param dataframe a dataframe or an object that can be coerced to a dataframe
 #' @param confidence what confidence level for confidence intervals
 #' @param plottype bar or line (quoted)
+#' @param errorbar.display default "CI" (confidence interval), which type of
+#'   errorbar should be displayed around the mean point? Other options
+#'   include "SEM" (standard error of the mean) and "SD" (standard dev).
 #' @param PlotSave a logical indicating whether the user wants to save the plot
 #'  as a png file
 #' @param xlab,ylab Labels for `x` and `y` axis variables. If `NULL` (default),
@@ -101,6 +109,16 @@
 #'   (Default: `"scheffe"`).
 #' @param show.dots Logical that decides whether the individual data points
 #'   are displayed (Default: `FALSE`).
+#' @param package Name of package from which the palette is desired as string
+#'   or symbol.
+#' @param palette Name of palette as string or symbol.
+#' @param ggtheme A function, ggplot2 theme name. Default value is ggplot2::theme_bw().
+#'   Any of the ggplot2 themes, or themes from extension packages are allowed (e.g.,
+#'   hrbrthemes::theme_ipsum(), etc.).
+#' @param ggplot.component A ggplot component to be added to the plot prepared.
+#'   The default is NULL. The argument should be entered as a function.
+#'   If the given function has an argument axes.range.restrict and if it has been set
+#'   to TRUE, the added ggplot component might not work as expected.
 #' @return A list with 5 elements which is returned invisibly. These items
 #'   are always sent to the console for display but for user convenience
 #'   the function also returns a named list with the following items
@@ -124,35 +142,16 @@
 #'   overlay.type = "box",
 #'   mean.label = TRUE
 #' )
-#' Plot2WayANOVA(mpg ~ am * vs, mtcars, confidence = .99)
-#'
-#' # Create a new dataset
-#' library(dplyr)
+#' 
 #' library(ggplot2)
-#' library(stringi)
-#' newmpg <- mpg %>%
-#'   filter(cyl != 5) %>%
-#'   mutate(am = stringi::stri_extract(trans, regex = "auto|manual"))
-#' Plot2WayANOVA(
-#'   formula = hwy ~ am * cyl,
-#'   dataframe = newmpg,
-#'   ylab = "Highway mileage",
-#'   xlab = "Transmission type",
-#'   plottype = "line",
-#'   offset.style = "wide",
-#'   overlay.type = "box",
-#'   mean.label = TRUE,
-#'   mean.shape = 20,
-#'   mean.size = 5,
-#'   mean.label.size = 5,
-#'   show.dots = TRUE
-#' )
+#' Plot2WayANOVA(mpg ~ am * vs, mtcars, confidence = .99,
+#'   ggplot.component = theme(axis.text.x = element_text(size=13, color="darkred")))
 #' @importFrom dplyr group_by summarise %>% n select filter
 #' @import ggplot2
 #' @import rlang
 #' @importFrom methods is
 #' @importFrom stats anova aov lm pf qt replications sd symnum residuals shapiro.test
-#' @importFrom dplyr as_tibble
+#' @importFrom dplyr as_tibble case_when
 #' @importFrom car leveneTest Anova
 #' @importFrom sjstats anova_stats
 #' @importFrom broom glance
@@ -163,6 +162,7 @@ Plot2WayANOVA <- function(formula,
                           dataframe = NULL,
                           confidence = .95,
                           plottype = "line",
+                          errorbar.display = "CI",
                           xlab = NULL,
                           ylab = NULL,
                           title = NULL,
@@ -180,13 +180,19 @@ Plot2WayANOVA <- function(formula,
                           overlay.type = NULL,
                           posthoc.method = "scheffe",
                           show.dots = FALSE,
-                          PlotSave = FALSE) {
+                          PlotSave = FALSE,
+                          ggtheme = ggplot2::theme_bw(),
+                          package = "RColorBrewer",
+                          palette = "Dark2",
+                          ggplot.component = NULL) {
 
   # -------- error checking ----------------
   if (!requireNamespace("ggplot2")) {
     stop("Can't continue can't load ggplot2")
   }
-  theme_set(theme_bw())
+  # set default theme 
+  ggplot2::theme_set(ggtheme)
+  
   if (!requireNamespace("dplyr")) {
     stop("Can't continue can't load dplyr")
   }
@@ -316,11 +322,27 @@ Plot2WayANOVA <- function(formula,
       TheSD = sd(!!sym(depvar), na.rm = TRUE),
       TheSEM = sd(!!sym(depvar), na.rm = TRUE) / sqrt(n()),
       CIMuliplier = qt(confidence / 2 + .5, n() - 1),
-      LowerBound = TheMean - TheSEM * CIMuliplier,
-      UpperBound = TheMean + TheSEM * CIMuliplier,
+      LowerBoundCI = TheMean - TheSEM * CIMuliplier,
+      UpperBoundCI = TheMean + TheSEM * CIMuliplier,
+      LowerBoundSEM = TheMean - TheSEM,
+      UpperBoundSEM = TheMean + TheSEM,
+      LowerBoundSD = TheMean - TheSD,
+      UpperBoundSD = TheMean + TheSD,
       N = n()
+    ) %>%
+    mutate(
+      LowerBound = case_when(
+        errorbar.display == "SD" ~ LowerBoundSD,
+        errorbar.display == "SEM" ~ LowerBoundSEM,
+        errorbar.display == "CI" ~ LowerBoundCI,
+        TRUE ~ LowerBoundCI),
+      UpperBound = case_when(
+        errorbar.display == "SD" ~ UpperBoundSD,
+        errorbar.display == "SEM" ~ UpperBoundSEM,
+        errorbar.display == "CI" ~ UpperBoundCI,
+        TRUE ~ UpperBoundCI)
     )
-
+  
   # -------- Run tests and procedures ----------------
 
   # run analysis of variance
@@ -359,11 +381,16 @@ Plot2WayANOVA <- function(formula,
   cipercent <- round(confidence * 100, 2)
   # if `title` is not provided, use this generic
   if (is.null(title)) {
-    title <- bquote(
-      "Group means with" ~ .(cipercent) * "% confidence intervals"
-    )
+    if (errorbar.display == "CI") {
+      title <- bquote(
+        "Group means with" ~ .(cipercent) * "% confidence intervals")
+    } else if (errorbar.display == "SEM") {
+      title <- "Group means with standard error of the mean"
+    } else if (errorbar.display == "SD") {
+      title <- "Group means with standard deviation"
+    }
   }
-
+  
   # compute CI's for R squared using Olkin and Finn's approximation
   denominator <- (nrow(dataframe)^2 - 1) * (3 + nrow(dataframe))
   numerator <- (4 * model_summary$r.squared) * ((1 - model_summary$r.squared)^2) * (nrow(dataframe) - 2 - 1)^2
@@ -611,8 +638,11 @@ Plot2WayANOVA <- function(formula,
     print(SWTest)
   }
   
-  # -------- Print the plot itself ----------------
-
+  ### -----  adding optional ggplot.component ----------
+  p <- p + ggplot.component
+  
+  #### -------- Print the plot itself ----------------
+  
   message("\nInteraction graph plotted...")
   print(p)